refactored and removed dead code
